smo-ui/pyproject.toml
[project]
name = "smo-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[all]>=0.116.1",
    "granian[reload]>=2.5.0",
    "pyyaml>=6.0.2",
    "smo-core",
    "sqlalchemy>=2.0.42",
]

[dependency-groups]
dev = [
    "coverage>=7.10.2",
    "isort>=6.0.1",
    "nox>=2025.5.1",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]


[build-system]
requires = ["uv-build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "smo_ui"
module-root = "src"
